+++ test/lua/test_veafCombatMission.lua
@@ -16,6 +16,7 @@ dofile(src .. "/veaf.lua")
 dofile(src .. "/veafScheduler.lua")
 dofile(src .. "/veafMath.lua")
 dofile(src .. "/veafGeo.lua")
+dofile(src .. "/veafMissionDb.lua")
 dofile(src .. "/veafI18n.lua")
 dofile(src .. "/veafCombatMission.lua")