+++ test/lua/test_veafCommands.lua
@@ -20,6 +20,7 @@ dofile(src .. "/veaf.lua")
 dofile(src .. "/veafScheduler.lua")
 dofile(src .. "/veafMath.lua")
 dofile(src .. "/veafGeo.lua")
+dofile(src .. "/veafMissionDb.lua")
 dofile(src .. "/veafMarkers.lua")
 dofile(src .. "/veafCommands.lua")