+++ test/lua/test_veafScheduler.lua
@@ -25,6 +25,7 @@ dofile(_base .. "/../../src/scripts/veaf/veaf.lua")
 dofile(_base .. "/../../src/scripts/veaf/veafScheduler.lua")
 dofile(_base .. "/../../src/scripts/veaf/veafMath.lua")
 dofile(_base .. "/../../src/scripts/veaf/veafGeo.lua")
+dofile(_base .. "/../../src/scripts/veaf/veafMissionDb.lua")
 
 TestVeafScheduler = {}