+++ test/lua/test_veafSkynetIadsHelper.lua
@@ -7,6 +7,7 @@ dofile(src .. "/veaf.lua")
 dofile(src .. "/veafScheduler.lua")
 dofile(src .. "/veafMath.lua")
 dofile(src .. "/veafGeo.lua")
+dofile(src .. "/veafMissionDb.lua")
 dofile(src .. "/veafSkynetIadsHelper.lua")
 
 -- ---------------------------------------------------------------------------